+++ TongSanTris/tetris/Tetris.py
@@ -2,7 +2,7 @@ import pygame, sys, time
 from pygame.locals import *
 from Board import *
 import random
-from ai import Ai
+# from ai import Ai
 from variable import Var
 
 
@@ -19,7 +19,7 @@ class Tetris:
         # self.interval = Var.keyboard_interval
         self.game = False
 
-        random.seed(Var.ai_random_seed)
+        #random.seed(Var.ai_random_seed)
         self.max_height = Var.display_max_height
         self.min_height = Var.display_min_height
         font_resize = Var.font_resize
@@ -48,85 +48,84 @@ class Tetris:
                 Var.base_bgm.stop()
                 Var.ai_bgm.stop()
 
-    def ai_new_stone(self):
-        self.stone = self.next_stone[:]
-        self.next_stone = Var.ai_tetris_shapes[
-            random.randint(Var.ai_block_choice_start, Var.ai_block_choice_end)]  # 다음 블럭 랜덤으로 고르기 0~6 사이의 랜덤 숫자를 통해 고르기
-        self.stone_x = int((self.board.width - len(
-            Var.piece_length(self.stone))) * Var.ai_stone_start_x_rate)  # self.width 기준 스톤의 위치 x
-        self.stone_y = Var.ai_stone_start_y
-        if self.board.ai_check_collision(self.ai_board, self.stone, (self.stone_x, self.stone_y)):
-            self.gameover = True  # 블럭이 부딪히는 판단, 새점로 생성한 블럭이 벽에 부딪히면은 게임 종료
+    # def ai_new_stone(self):
+        # self.stone = self.next_stone[:]
+        # self.next_stone = Var.ai_tetris_shapes[
+            # random.randint(Var.ai_block_choice_start, Var.ai_block_choice_end)]  # 다음 블럭 랜덤으로 고르기 0~6 사이의 랜덤 숫자를 통해 고르기
+        # self.stone_x = int((self.board.width - len(
+            # Var.piece_length(self.stone))) * Var.ai_stone_start_x_rate)  # self.width 기준 스톤의 위치 x
+        # self.stone_y = Var.ai_stone_start_y
+        # if self.board.ai_check_collision(self.ai_board, self.stone, (self.stone_x, self.stone_y)):
+            # self.gameover = True  # 블럭이 부딪히는 판단, 새점로 생성한 블럭이 벽에 부딪히면은 게임 종료
 
-    def ai_init_game(self):
-        self.ai_board = [[Var.board_empty_state for x in range(self.board.width)] for y in
-                         range(self.board.height)]  # 새로운 게임 보드 생성
-        self.ai_new_stone()  # 새로운 블럭 생성
-        self.ai_score = Var.initial_score  # 시작 스코어
-        self.ai_lines = Var.ai_line_reset  # 지운 라인의 개수
+    # def ai_init_game(self):
+        # self.ai_board = [[Var.board_empty_state for x in range(self.board.width)] for y in
+                        #  range(self.board.height)]  # 새로운 게임 보드 생성
+        # self.ai_new_stone()  # 새로운 블럭 생성
+        # self.ai_score = Var.initial_score  # 시작 스코어
+        # self.ai_lines = Var.ai_line_reset  # 지운 라인의 개수
 
-    def ai_add_cl_lines(self, n):
-        self.ai_lines += n  # 지운 개수 추가하기
-        self.ai_score += Var.ai_linescores[
-                             n] * self.board.level * Var.ai_score_weight  # 점수  = 원래 점수 + 한번에지운 라인개수에 해당하는 점수 * 레벨
+    # def ai_add_cl_lines(self, n):
+        # self.ai_lines += n  # 지운 개수 추가하기
+        # self.ai_score += Var.ai_linescores[
+    #                     n] * self.board.level * Var.ai_score_weight  # 점수  = 원래 점수 + 한번에지운 라인개수에 해당하는 점수 * 레벨
 
         # 블럭을 delta_x 만큼 움직이기
 
-    def ai_move(self, delta_x):
-        if not self.gameover and not self.paused:  # 게임 종료, 정지 상태가 아니라면
-            new_x = self.stone_x + delta_x  # 새로운 x 좌표는   기존의 stone의 x좌표 + 이동좌표수
-            if new_x < Var.board_start_x:  # 새로운 좌표가 0보다 작다면
-                new_x = Var.board_start_x  # 새로운 좌표는 0 ( 변경 없음 )
-            if new_x > self.board.width - len(Var.piece_length(self.stone)):  # 새로운 좌표 > 열의개수(10) - 블럭의 x축 길이
-                new_x = self.board.width - len(Var.piece_length(self.stone))  # 이동 불가 (변경 없음)
-            if not self.board.ai_check_collision(self.ai_board, self.stone, (new_x, self.stone_y)):  # 벽과 부딪히지 않는 다면
-                self.stone_x = new_x  # 새로운 좌표로 이동
+    # def ai_move(self, delta_x):
+        # if not self.gameover and not self.paused:  # 게임 종료, 정지 상태가 아니라면
+            # new_x = self.stone_x + delta_x  # 새로운 x 좌표는   기존의 stone의 x좌표 + 이동좌표수
+            # if new_x < Var.board_start_x:  # 새로운 좌표가 0보다 작다면
+            #    new_x = Var.board_start_x  # 새로운 좌표는 0 ( 변경 없음 )
+            # if new_x > self.board.width - len(Var.piece_length(self.stone)):  # 새로운 좌표 > 열의개수(10) - 블럭의 x축 길이
+            #     new_x = self.board.width - len(Var.piece_length(self.stone))  # 이동 불가 (변경 없음)
+            # if not self.board.ai_check_collision(self.ai_board, self.stone, (new_x, self.stone_y)):  # 벽과 부딪히지 않는 다면
+            #     self.stone_x = new_x  # 새로운 좌표로 이동
 
-    def ai_drop(self, manual):
-        if not self.gameover and not self.paused:  # 게임 종료, 정지 상ef태가 아니라면
-            self.ai_score += Var.ai_blockdown_score_per if manual else Var.ai_no_blockdown_score_per  # 내릴 수 있다면  스코어+1, 내릴수 없다면 +0
-            self.stone_y += Var.y_move_scale  # y축 좌표 +1
-            if self.board.ai_check_collision(self.ai_board, self.stone,
-                                             (self.stone_x, self.stone_y)):  # 벽에 부딪히지 않는 경우에는
-                self.ai_board = self.board.join_matrixes(self.ai_board, self.stone,
-                                                         (self.stone_x,
-                                                          self.stone_y))  # 새로운 보드는   블럭의 새로운 좌표를 포합한 보드로 갱신
-                self.ai_score += Var.ai_blockdown_score_per  # 블럭이 밑에 내려가면 점수 추가
-                self.ai_new_stone()  # 새로운 블럭 생성하기
-                cleared_rows = Var.ai_line_reset  # 지운 행의 개수 초기화
+    # def ai_drop(self, manual):
+    #     if not self.gameover and not self.paused:  # 게임 종료, 정지 상ef태가 아니라면
+    #        self.ai_score += Var.ai_blockdown_score_per if manual else Var.ai_no_blockdown_score_per  # 내릴 수 있다면  스코어+1, 내릴수 없다면 +0
+    #        self.stone_y += Var.y_move_scale  # y축 좌표 +1
+    #        if self.board.ai_check_collision(self.ai_board, self.stone,
+    #                                         (self.stone_x, self.stone_y)):  # 벽에 부딪히지 않는 경우에는
+    #            self.ai_board = self.board.join_matrixes(self.ai_board, self.stone,
+    #                                                     (self.stone_x,
+    #                                                      self.stone_y))  # 새로운 보드는   블럭의 새로운 좌표를 포합한 보드로 갱신
+    #            self.ai_score += Var.ai_blockdown_score_per  # 블럭이 밑에 내려가면 점수 추가
+    #            self.ai_new_stone()  # 새로운 블럭 생성하기
+    #            cleared_rows = Var.ai_line_reset  # 지운 행의 개수 초기화
 
-                for i, row in enumerate(self.ai_board):  # 모든 행마다 검사    i는 행의 번호
-                    if Var.board_empty_state not in row:  # 행에 빈공간이 없다면 (0은 빈공간의 의미한다.)
-                        self.ai_board = self.board.ai_remove_row(self.ai_board, i)  # ai보드를 i행을 지운 보드로 업데이트
-                        cleared_rows += Var.ai_lineclear_per  # 지운개수 +1  # 한번에 지운 개수에 해당하는 만큼    지운라인개수, 점수, 레벨 변경
-                self.ai_add_cl_lines(cleared_rows)  # 한번에 지운 개수에 해당하는 만큼    지운라인개수, 점수, 레벨 변경
-                return True  # 게임 종료 상태가 아니 었다면,  true 반환
-        return False  # 게임이 정지, 종료 상태 였다면 false 반환
+    #            for i, row in enumerate(self.ai_board):  # 모든 행마다 검사    i는 행의 번호
+    #                if Var.board_empty_state not in row:  # 행에 빈공간이 없다면 (0은 빈공간의 의미한다.)
+    #                    self.ai_board = self.board.ai_remove_row(self.ai_board, i)  # ai보드를 i행을 지운 보드로 업데이트
+    #            self.ai_add_cl_lines(cleared_rows)  # 한번에 지운 개수에 해당하는 만큼    지운라인개수, 점수, 레벨 변경
+    #            return True  # 게임 종료 상태가 아니 었다면,  true 반환
+    #    return False  # 게임이 정지, 종료 상태 였다면 false 반환
 
-    def rotate_stone(self):
-        if not self.gameover and not self.paused:
-            ai_new_stone = self.board.ai_rotate_clockwise(self.stone)
-            if not self.board.ai_check_collision(self.ai_board,
-                                                 ai_new_stone,
-                                                 (self.stone_x, self.stone_y)):
-                self.stone = ai_new_stone
+    # def rotate_stone(self):
+    #    if not self.gameover and not self.paused:
+    #        ai_new_stone = self.board.ai_rotate_clockwise(self.stone)
+    #        if not self.board.ai_check_collision(self.ai_board,
+    #                                             ai_new_stone,
+    #                                             (self.stone_x, self.stone_y)):
+    #            self.stone = ai_new_stone
 
-    def ai_start_game(self):
-        if self.gameover:  # 게임이 종료된 상태라면
-            self.ai_init_game()  # 초기화 진행하기
-            self.gameover = False  # 게임 종료는 false로 바꿔주기
+    # def ai_start_game(self):
+    #    if self.gameover:  # 게임이 종료된 상태라면
+    #        self.ai_init_game()  # 초기화 진행하기
+    #        self.gameover = False  # 게임 종료는 false로 바꿔주기
 
-    def ai_executes_moves(self, ai_moves):
-        ai_key_actions = {
-            'LEFT': lambda: self.ai_move(-Var.x_move_scale),
-            'RIGHT': lambda: self.ai_move(+Var.x_move_scale),
-            'DOWN': lambda: self.ai_drop(True),
-            'UP': self.rotate_stone,
-            'SPACE': self.ai_start_game
-        }
+    # def ai_executes_moves(self, ai_moves):
+    #    ai_key_actions = {
+    #        'LEFT': lambda: self.ai_move(-Var.x_move_scale),
+    #        'RIGHT': lambda: self.ai_move(+Var.x_move_scale),
+    #        'DOWN': lambda: self.ai_drop(True),
+    #        'UP': self.rotate_stone,
+    #        'SPACE': self.ai_start_game
+    #    }
         # 받아온 moves에 저장되어 있는 동작 수행하기  -- (ai에서 학습된 것을 통해 결정 되는 부분)
-        for ai_action in ai_moves:
-            ai_key_actions[ai_action]()
+    #    for ai_action in ai_moves:
+    #        ai_key_actions[ai_action]()
 
     def handle_key2(self, event_key, mode): # 투핸즈 모드 게임 조작 키 설정
         if event_key == K_s:
@@ -314,9 +313,9 @@ class Tetris:
                     if self.mode == 'two':
                         self.board.drop_piece2()
 
-                elif event.type == Var.ai_event:
-                    if self.mode == 'ai':
-                        self.ai_drop(False)
+        #        elif event.type == Var.ai_event:
+        #            if self.mode == 'ai':
+        #                self.ai_drop(False)
 
                 # 화면 크기 조절해 보기
                 elif event.type == VIDEORESIZE:
